TF:
- FIX: fix/surpress/cast-away (take your pick) several warnings in the CUDA code in x86_64 mode. Usually about size_t to unsigned int conversion.

+++ src/Common/CUDA/cudaResampleImageFilter.cuh
@@ -65,7 +65,7 @@ namespace cuda
 		cudaChannelFormatDesc m_channelDescCoeff;
 		int                   m_Device;
 
-		size_t                m_MaxnrOfVoxelsPerIteration;
+		unsigned int          m_MaxnrOfVoxelsPerIteration;
 		TInternalImageType* inputImage;
 
 	#if defined(__CUDACC__)
